[ANDROSDK-2129] refactor ItemFilterTableInfo into FilterOperatorsColumns
